+++ lib/ui/loan/vault_detail.dart
@@ -232,7 +232,7 @@ class _VaultDetailScreen extends State<VaultDetailScreen> with SingleTickerProvi
   _buildCollateralEntry(LoanVaultAmount amount) {
     var token = _tokens.firstWhere((element) => amount.symbol == element.token.symbol, orElse: () => null);
     double price = amount.activePrice != null ? amount.activePrice.active.amount : 0;
-    int factor = token != null ? int.tryParse(token.factor) : 0;
+    double factor = token != null ? double.tryParse(token.factor) : 0;
 
     return Card(
         child: Padding(
